pkg/monitor: fix build, clippy, and lint warnings
Make `cargo build` and `cargo clippy` warning-free on the vendored monitor
source: mechanical clippy fixes (ptr_arg, matches!, vec![], borrowed_box,
infallible conversions, ...), delete a small amount of clearly-dead code, and
apply targeted #[allow(...)] for intended-but-unused API (UI widget/trait
surface) and the hand-maintained type/spec mirrors (eve_types, tcg).

Also clear the project linter findings reported by yetus: trailing whitespace,
codespell typos, a README top-level heading and trailing newline, and hadolint
findings in the Dockerfile. No behavior change; the crate tests still pass.

(cherry picked from commit 0d14b2c2fffab7bf062d43607572b10fbe353c9d)

+++ pkg/monitor/src/model/device/dmesg.rs
@@ -63,7 +63,7 @@ impl DmesgViewer {
         match key.code {
             KeyCode::Down => {
                 self.buffer_index = cmp::min(
-                    self.buffer_index + 1 as usize,
+                    self.buffer_index + 1_usize,
                     self.buffer_len - self.lines_per_page as usize,
                 );
             }
